organize extension files in textext/ subfolder
Avoid pollution of inkscape extension folder
