+++ extension/textext.inx
@@ -1,16 +1,16 @@
 <inkscape-extension>
   <_name>Tex Text</_name>
   <id>org.ekips.filter.textext</id>
-  <dependency type="executable" location="extensions">textext.py</dependency>
   <dependency type="executable" location="extensions">inkex.py</dependency>
-  <dependency type="executable" location="extensions">asktext.py</dependency>
-  <dependency type="executable" location="extensions">typesetter.py</dependency>
-  <dependency type="executable" location="extensions">latexlogparser.py</dependency>
-  <dependency type="executable" location="extensions">win_app_paths.py</dependency>
+  <dependency type="executable" location="extensions">textext/__init__.py</dependency>
+  <dependency type="executable" location="extensions">textext/asktext.py</dependency>
+  <dependency type="executable" location="extensions">textext/typesetter.py</dependency>
+  <dependency type="executable" location="extensions">textext/latexlogparser.py</dependency>
+  <dependency type="executable" location="extensions">textext/win_app_paths.py</dependency>
   <effect>
     <object-type>all</object-type>
   </effect>
   <script>
-    <command reldir="extensions" interpreter="python">textext.py</command>
+    <command reldir="extensions" interpreter="python">textext/__init__.py</command>
   </script>
 </inkscape-extension>
